refactor(rules): split rules.ts god file into rule definition modules
rules/rules.ts had grown to ~830 lines, mixing ~14 inline rule definitions and
a large wiki channel-name-resolution helper block in with the ruleList
aggregation. Move the inline rules into focused modules under rules/definitions/:

- contentCreationRules.ts  canCreate{Channel,Discussion,Event,Comment} + types
- wikiRules.ts             wiki helpers + evaluateCanEdit/Delete* + 3 rules
- votingRules.ts           canUpvote{Comment,Discussion}
- serverRules.ts           isAdmin, canUploadFile, canGiveFeedback, ...

rules.ts is now a ~140-line aggregator that wires every rule into `ruleList`
(default export, unchanged) and re-exports the evaluate* helpers and arg/item
types that tests and the validation modules import from "./rules". No logic
changes.

tsc clean; 565/565 unit tests pass (incl. canCreateChannelRule and
canEditWikiPagesRule which exercise the re-exported evaluate* functions).
