+++ rules/rules.ts
@@ -1,10 +1,12 @@
-import { rule } from "graphql-shield";
-import type { GraphQLResolveInfo } from "graphql";
-import type { GraphQLContext } from "../types/context.js";
-import { ERROR_MESSAGES } from "./errorMessages.js";
+// Aggregates every graphql-shield rule into the `ruleList` consumed by
+// permissions.ts. The rule *definitions* live in focused modules:
+//   - permission/isOwner.js            ownership/authorship rules
+//   - permission/*                     channel/server permission rules
+//   - validation/*                     input validation rules
+//   - definitions/*                    rules previously inlined in this file
+// This file only wires them together and re-exports the handful of helpers and
+// types that tests and validation modules import from "./rules".
 import { isAuthenticatedAndVerified, isAuthenticated } from "./permission/userDataHelperFunctions.js";
-import { hasServerPermission } from "./permission/hasServerPermission.js";
-import { hasServerModPermission } from "./permission/hasServerModPermission.js";
 import {
   isChannelOwner,
   isAccountOwner,
@@ -17,11 +19,7 @@ import {
   isIssueAuthor,
   issueIsNotLocked,
 } from "./permission/isOwner.js";
-import {
-  hasChannelPermission,
-} from "./permission/hasChannelPermission.js";
-import { checkChannelPermissions } from "./permission/hasChannelPermission.js";
-import { checkChannelModPermissions, ModChannelPermission } from "./permission/hasChannelModPermission.js";
+import { hasChannelPermission } from "./permission/hasChannelPermission.js";
 import { canArchiveAndUnarchiveDiscussion } from "./permission/canArchiveAndUnarchiveDiscussion.js";
 import { canArchiveAndUnarchiveEvent } from "./permission/canArchiveAndUnarchiveEvent.js";
 import { canArchiveAndUnarchiveComment } from "./permission/canArchiveAndUnarchiveComment.js";
@@ -34,19 +32,6 @@ import { canReport } from "./permission/canReport.js";
 import { canSuspendAndUnsuspendUser } from "./permission/canSuspendAndUnsuspendUser.js";
 import { canBecomeForumAdmin } from "./permission/canBecomeForumAdmin.js";
 import { canLockChannel } from "./permission/canLockChannel.js";
-import {
-  CommentCreateInput,
-  DiscussionCreateInput,
-  EventCreateInput,
-  EventUpdateInput,
-  MutationCreateWikiPagesArgs,
-  MutationDeleteWikiPagesArgs,
-  MutationUpdateChannelsArgs,
-  MutationUpdateWikiPagesArgs,
-  WikiPageChildPagesUpdateFieldInput,
-  WikiPageCreateInput,
-  WikiPageWhere,
-} from "../src/generated/graphql.js";
 import {
   createDiscussionInputIsValid,
   updateDiscussionInputIsValid,
@@ -68,713 +53,40 @@ import {
   updateDownloadableFileInputIsValid,
 } from "./validation/downloadableFileIsValid.js";
 import { updateUserInputIsValid } from "./validation/userIsValid.js";
-import { getServerScopedMembership } from "./permission/getServerScopedMembership.js";
-
-export async function evaluateCanCreateChannelRule(ctx: GraphQLContext) {
-  const hasPermissionToCreateChannels = await hasServerPermission(
-    "canCreateChannel",
-    ctx
-  );
-
-  if (hasPermissionToCreateChannels instanceof Error) {
-    return hasPermissionToCreateChannels;
-  }
-
-  return true;
-}
-
-const canCreateChannel = rule({ cache: "contextual" })(
-  async (parent: unknown, args: unknown, ctx: GraphQLContext, info: GraphQLResolveInfo) =>
-    evaluateCanCreateChannelRule(ctx)
-);
-
-export type CreateDiscussionItem = {
-  discussionCreateInput: DiscussionCreateInput;
-  channelConnections: string[];
-};
-
-export type CanCreateDiscussionArgs = {
-  input: CreateDiscussionItem[];
-};
-
-export type CanUpdateDiscussionArgs = {
-  discussionUpdateInput: DiscussionCreateInput;
-  channelConnections: string[];
-};
-
-export const canCreateDiscussion = rule({ cache: "contextual" })(
-  async (parent: unknown, args: CanCreateDiscussionArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const inputItems = args.input;
-    for (let i = 0; i < inputItems.length; i++) {
-      const item = inputItems[i];
-      const { channelConnections } = item;
-
-      const channelPermissions = await checkChannelPermissions({
-        channelConnections,
-        context: ctx,
-        permissionCheck: "canCreateDiscussion",
-      });
-
-      if (channelPermissions instanceof Error) {
-        return channelPermissions;
-      }
-    }
-    return true;
-  }
-);
-
-export type SingleEventInput = {
-  eventCreateInput: EventCreateInput;
-  channelConnections: string[];
-}
-
-export type CanCreateEventArgs = {
-  input: SingleEventInput[];
-};
-
-export const canCreateEvent = rule({ cache: "contextual" })(
-  async (parent: unknown, args: CanCreateEventArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const dedupedChannelConnections = args.input.map((item) => item.channelConnections);
-    const channelConnections = [...new Set(dedupedChannelConnections)];
-    const flattenedChannelConnections = channelConnections.flat();
-    
-    return checkChannelPermissions({
-      channelConnections: flattenedChannelConnections,
-      context: ctx,
-      permissionCheck: "canCreateEvent",
-    });
-  }
-);
-
-type CanCreateCommentArgs = {
-  input: CommentCreateInput[];
-};
-
-export const canCreateComment = rule({ cache: "contextual" })(
-  async (parent: unknown, args: CanCreateCommentArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const { input } = args;
-    const firstItemInInput = input[0];
-
-    if (!firstItemInInput) {
-      throw new Error("No comment create input found.");
-    }
-
-    const { 
-      DiscussionChannel, 
-      Event,
-      GivesFeedbackOnEvent,
-      GivesFeedbackOnDiscussion,
-      GivesFeedbackOnComment,
-      Channel,
-    } = firstItemInInput;
-
-    // Throw an error if no Channel is provided; all comments must be in the context of a channel.
-    if (!Channel || !Channel.connect?.where?.node?.uniqueName) {
-      throw new Error("Comment must be connected to a Channel.");
-    }
-
-    let channelName = '';
-    
-    // Determine if this is a feedback comment
-    const isFeedbackComment = !!(GivesFeedbackOnEvent || GivesFeedbackOnDiscussion || GivesFeedbackOnComment);
-
-    if (DiscussionChannel){
-      const discussionChannelId = DiscussionChannel.connect?.where?.node?.id;
-
-      if (!discussionChannelId) {
-        throw new Error("No discussion channel ID found.");
-      }
-  
-      // Look up the channelUniqueName from the discussion channel ID.
-      const discussionChannelModel = ctx.ogm.model("DiscussionChannel");
-      const discussionChannel = await discussionChannelModel.find({
-        where: { id: discussionChannelId },
-        selectionSet: `{ channelUniqueName }`,
-      });
-
-      if (!discussionChannel || !discussionChannel[0]) {
-        throw new Error("No discussion channel found.");
-      }
-  
-      channelName = discussionChannel[0]?.channelUniqueName;
-    }
-
-    if (Event) {
-      const eventId = Event.connect?.where?.node?.id;
-
-      if (!eventId) {
-        throw new Error("No event ID found.");
-      }  
-
-      // Validate that the user has permission to comment on the event.
-      // The channel that they are posting in needs to match one of the
-      // channels that the event is connected to.
-      const eventChannelModel = ctx.ogm.model("EventChannel");
-      const event = await eventChannelModel.find({
-        where: { 
-          eventId,
-          channelUniqueName: Channel?.connect?.where?.node?.uniqueName
-        },
-        selectionSet: `{ id }`,
-      });
-
-      if (!event || !event[0]) {
-        throw new Error("Could not find the event submission in the given channel.");
-      }
-
-      channelName = Channel?.connect?.where?.node?.uniqueName
-    }
-
-    if (GivesFeedbackOnEvent) {
-      const eventId = GivesFeedbackOnEvent.connect?.where?.node?.id;
-
-      if (!eventId) {
-        throw new Error("No event ID found.");
-      }
-
-      // Validate that the user has permission to comment on the event.
-      // The channel that they are posting in needs to match one of the
-      // channels that the event is connected to.
-      const eventChannelModel = ctx.ogm.model("EventChannel");
-      const event = await eventChannelModel.find({
-        where: { 
-          eventId,
-          channelUniqueName: Channel?.connect?.where?.node?.uniqueName
-        },
-        selectionSet: `{ id }`,
-      });
-
-      if (!event || !event[0]) {
-        throw new Error("Could not find the event submission in the given channel.");
-      }
-
-      channelName = Channel?.connect?.where?.node?.uniqueName
-    }
-
-    if (GivesFeedbackOnDiscussion) {
-      // GivesFeedbackOnDiscussion is of type Discussion which doesn't directly have
-      // the channel on it, so we use the DiscussionChannel to get the channel name.
-      const discussionChannelId = DiscussionChannel?.connect?.where?.node?.id;
-
-      if (!discussionChannelId) {
-        throw new Error("No discussion channel ID found.");
-      }
-  
-      // Look up the channelUniqueName from the discussion channel ID.
-      const discussionChannelModel = ctx.ogm.model("DiscussionChannel");
-      const discussionChannel = await discussionChannelModel.find({
-        where: { id: discussionChannelId },
-        selectionSet: `{ channelUniqueName }`,
-      });
-  
-      if (!discussionChannel || !discussionChannel[0]) {
-        throw new Error("No discussion channel found.");
-      }
-  
-      channelName = discussionChannel[0]?.channelUniqueName;
-    }
-
-    if (GivesFeedbackOnComment) {
-      const commentId = GivesFeedbackOnComment?.connect?.where?.node?.id;
-
-      if (!commentId) {
-        throw new Error("No comment ID found.");
-      }
-  
-      channelName = Channel?.connect?.where?.node?.uniqueName;
-    }
-    
-    if (!channelName) {
-      throw new Error("No channel name found.");
-    }
-
-    // Different permission checking based on whether it's a feedback comment
-    if (isFeedbackComment) {
-      // For feedback comments, check mod permissions with canGiveFeedback
-      return checkChannelModPermissions({
-        channelConnections: [channelName],
-        context: ctx,
-        permissionCheck: ModChannelPermission.canGiveFeedback
-      });
-    } else {
-      // For regular comments, check regular permissions
-      return checkChannelPermissions({
-        channelConnections: [channelName],
-        context: ctx,
-        permissionCheck: "canCreateComment",
-      });
-    }
-  }
-);
-
-const isAdmin = rule({ cache: "contextual" })(
-  async (parent: unknown, args: unknown, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const membership = await getServerScopedMembership(ctx);
-    return membership.isServerAdmin;
-  }
-);
-
-const canUploadFile = rule({ cache: "contextual" })(
-  async (parent: unknown, args: unknown, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const permissionResult = await hasServerPermission(
-      "canUploadFile",
-      ctx
-    );
-
-    if (!permissionResult) {
-      return false;
-    }
-
-    if (permissionResult instanceof Error) {
-      return permissionResult;
-    }
-
-    return true;
-  }
-);
-
-type WikiPageLookup = {
-  id?: string | null;
-  channelUniqueName?: string | null;
-  OriginalAuthor?: {
-    username?: string | null;
-  } | null;
-  ChildPages?: Array<{
-    id?: string | null;
-  }> | null;
-};
-
-type WikiChannelPreference = {
-  uniqueName?: string | null;
-  wikiEnabled?: boolean | null;
-};
-
-type WikiPageMutationArgs = Partial<
-  MutationCreateWikiPagesArgs &
-    MutationUpdateWikiPagesArgs &
-    MutationDeleteWikiPagesArgs
->;
-
-type WikiPageChildPagesUpdate =
-  | WikiPageChildPagesUpdateFieldInput
-  | {
-      create?: Array<{ node?: Pick<WikiPageCreateInput, "channelUniqueName"> }>;
-    };
-
-function addChannelName(
-  channelNames: Set<string>,
-  channelName?: string | null
-) {
-  if (channelName) {
-    channelNames.add(channelName);
-  }
-}
-
-function hasWikiPageWhere(where?: WikiPageWhere | null) {
-  return !!where && Object.keys(where).length > 0;
-}
-
-async function getWikiPagesByWhere(where: WikiPageWhere | null, ctx: GraphQLContext) {
-  if (!hasWikiPageWhere(where)) {
-    return [];
-  }
-
-  const WikiPage = ctx.ogm.model("WikiPage");
-  return (await WikiPage.find({
-    where: where ?? undefined,
-    selectionSet: `{
-      id
-      channelUniqueName
-      OriginalAuthor {
-        username
-      }
-      ChildPages {
-        id
-      }
-    }`,
-  })) as WikiPageLookup[];
-}
-
-function collectWikiPageChannelNames(wikiPages: WikiPageLookup[]) {
-  const channelNames = new Set<string>();
-
-  wikiPages.forEach((wikiPage) => {
-    addChannelName(channelNames, wikiPage.channelUniqueName);
-  });
-
-  return Array.from(channelNames);
-}
-
-async function getWikiPageChannelNamesByWhere(
-  args: WikiPageMutationArgs,
-  ctx: GraphQLContext
-) {
-  const channelNames = new Set<string>();
-  const where = args?.where || null;
-
-  addChannelName(channelNames, where?.channelUniqueName);
-
-  const wikiPages = await getWikiPagesByWhere(where, ctx);
-  collectWikiPageChannelNames(wikiPages).forEach((channelName) => {
-    addChannelName(channelNames, channelName);
-  });
-
-  return Array.from(channelNames);
-}
-
-function getCreatedWikiPageChannelNames(
-  input?: MutationCreateWikiPagesArgs["input"]
-) {
-  const channelNames = new Set<string>();
-  const inputs = input || [];
-
-  inputs.forEach((item) => {
-    addChannelName(channelNames, item?.channelUniqueName);
-  });
-
-  return Array.from(channelNames);
-}
-
-function getChildPageCreateChannelNames(
-  childPages?: MutationUpdateWikiPagesArgs["update"] extends infer T
-    ? T extends { ChildPages?: infer U }
-      ? U
-      : never
-    : never
-) {
-  const updates = Array.isArray(childPages)
-    ? childPages
-    : childPages
-    ? ([childPages] as WikiPageChildPagesUpdate[])
-    : [];
-
-  return updates.flatMap(
-    (update: WikiPageChildPagesUpdate) =>
-      update.create
-        ?.map((item) => item?.node?.channelUniqueName)
-        .filter((channelName): channelName is string => !!channelName) || []
-  );
-}
-
-async function validateWikiChannelsEnabled(
-  channelConnections: string[],
-  ctx: GraphQLContext
-) {
-  const Channel = ctx.ogm.model("Channel");
-  const channelNames = Array.from(new Set(channelConnections.filter(Boolean)));
-
-  for (const channelName of channelNames) {
-    const channels = (await Channel.find({
-      where: { uniqueName: channelName },
-      selectionSet: `{
-        uniqueName
-        wikiEnabled
-      }`,
-    })) as WikiChannelPreference[];
-    const channel = channels[0];
-
-    if (channel?.wikiEnabled === false) {
-      return new Error(`Wiki is disabled in channel '${channelName}'.`);
-    }
-  }
-
-  return true;
-}
-
-export async function evaluateCanEditWikiPagesRule(
-  args: WikiPageMutationArgs,
-  ctx: GraphQLContext
-) {
-  const whereChannelNames = await getWikiPageChannelNamesByWhere(args, ctx);
-  const createdChildChannelNames = getChildPageCreateChannelNames(
-    args?.update?.ChildPages
-  );
-  const createdPageChannelNames = getCreatedWikiPageChannelNames(args?.input);
-  const channelConnections = Array.from(
-    new Set([
-      ...whereChannelNames,
-      ...createdChildChannelNames,
-      ...createdPageChannelNames,
-    ])
-  );
-  const wikiEnabledResult = await validateWikiChannelsEnabled(
-    channelConnections,
-    ctx
-  );
-
-  if (wikiEnabledResult instanceof Error) {
-    return wikiEnabledResult;
-  }
-
-  return checkChannelPermissions({
-    channelConnections,
-    context: ctx,
-    permissionCheck: "canUpdateChannel",
-  });
-}
-
-export async function evaluateCanEditWikiHomePageRule(
-  args: MutationUpdateChannelsArgs,
-  ctx: GraphQLContext
-) {
-  const wikiHomePageUpdate = args?.update?.WikiHomePage;
-
-  if (!wikiHomePageUpdate) {
-    return true;
-  }
-
-  const channelNames = new Set<string>();
-  addChannelName(channelNames, args?.where?.uniqueName);
-  addChannelName(
-    channelNames,
-    wikiHomePageUpdate?.create?.node?.channelUniqueName
-  );
-  addChannelName(
-    channelNames,
-    wikiHomePageUpdate?.update?.node?.channelUniqueName
-  );
-  const channelConnections = Array.from(channelNames);
-  const wikiEnabledResult = await validateWikiChannelsEnabled(
-    channelConnections,
-    ctx
-  );
-
-  if (wikiEnabledResult instanceof Error) {
-    return wikiEnabledResult;
-  }
-
-  return checkChannelPermissions({
-    channelConnections,
-    context: ctx,
-    permissionCheck: "canUpdateChannel",
-  });
-}
-
-export async function evaluateCanDeleteWikiPagesRule(
-  args: MutationDeleteWikiPagesArgs,
-  ctx: GraphQLContext,
-  checkModPermissions = checkChannelModPermissions
-) {
-  if (!hasWikiPageWhere(args?.where || null)) {
-    return new Error("No wiki page specified for this operation.");
-  }
-
-  const wikiPages = await getWikiPagesByWhere(args?.where || null, ctx);
-
-  if (!wikiPages.length) {
-    return new Error("Could not find the wiki page or its associated channel.");
-  }
-
-  const currentUsername = ctx.user?.username || null;
-  if (!currentUsername) {
-    return new Error(ERROR_MESSAGES.channel.notAuthenticated);
-  }
-
-  const pagesWithChildren = wikiPages.filter(
-    (wikiPage) => wikiPage.ChildPages?.length
-  );
-
-  if (pagesWithChildren.length) {
-    return new Error(
-      "Cannot delete a wiki page that has child pages. Delete child pages first."
-    );
-  }
-
-  const userOwnsAllPages = wikiPages.every(
-    (wikiPage) => wikiPage.OriginalAuthor?.username === currentUsername
-  );
-
-  if (userOwnsAllPages) {
-    return true;
-  }
-
-  const channelConnections = collectWikiPageChannelNames(wikiPages);
-
-  if (!channelConnections.length) {
-    return new Error("No channel specified for this operation.");
-  }
-
-  const permissionResult = await checkModPermissions({
-    channelConnections,
-    context: ctx,
-    permissionCheck: ModChannelPermission.canDeleteWiki,
-  });
-
-  if (permissionResult instanceof Error) {
-    return permissionResult;
-  }
-
-  return true;
-}
-
-const canEditWikiPages = rule({ cache: "contextual" })(
-  async (parent: unknown, args: WikiPageMutationArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) =>
-    evaluateCanEditWikiPagesRule(args, ctx)
-);
-
-const canDeleteWikiPages = rule({ cache: "contextual" })(
-  async (parent: unknown, args: MutationDeleteWikiPagesArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) =>
-    evaluateCanDeleteWikiPagesRule(args, ctx)
-);
-
-const canEditWikiHomePage = rule({ cache: "contextual" })(
-  async (parent: unknown, args: MutationUpdateChannelsArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) =>
-    evaluateCanEditWikiHomePageRule(args, ctx)
-);
-
-type CanUpvoteCommentArgs = {
-  commentId: string;
-  username: string;
-};
-
-const canUpvoteComment = rule({ cache: "contextual" })(
-  async (parent: unknown, args: CanUpvoteCommentArgs, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const CommentModel = ctx.ogm.model("Comment");
-
-    const { commentId, username } = args;
-
-    if (!commentId || !username) {
-      throw new Error("All arguments (commentId, username) are required");
-    }
-
-    const commentData = await CommentModel.find({
-      where: { id: commentId },
-      selectionSet: `{ 
-        id
-        DiscussionChannel {
-          channelUniqueName
-        }
-        Channel {
-          uniqueName
-        }
-      }`,
-    });
-
-    if (!commentData || !commentData[0]) {
-      throw new Error("No comment found.");
-    }
-
-    // Try to get channel name either from DiscussionChannel or directly from Channel
-    let channelThatCommentIsIn = commentData[0]?.DiscussionChannel?.channelUniqueName;
-    
-    // If not found in DiscussionChannel, try the direct Channel relationship
-    if (!channelThatCommentIsIn) {
-      channelThatCommentIsIn = commentData[0]?.Channel?.uniqueName;
-    }
-
-    if (!channelThatCommentIsIn) {
-      throw new Error("No channel found. Comment must be associated with a channel.");
-    }
-
-    const permissionResult = await hasChannelPermission({
-      permission: "canUpvoteComment",
-      channelName: channelThatCommentIsIn,
-      context: ctx,
-    });
-
-    if (!permissionResult) {
-      throw new Error(ERROR_MESSAGES.channel.noChannelPermission);
-    }
-
-    if (permissionResult instanceof Error) {
-      return permissionResult;
-    }
-
-    return true;
-  }
-);
-
-type CanUpvoteDiscussionChannelArgs = {
-  discussionChannelId: string;
-  username: string;
-};
-
-const canUpvoteDiscussion = rule({ cache: "contextual" })(
-  async (
-    parent: unknown,
-    args: CanUpvoteDiscussionChannelArgs,
-    ctx: GraphQLContext,
-    info: GraphQLResolveInfo
-  ) => {
-    const DiscussionChannelModel = ctx.ogm.model("DiscussionChannel");
-
-    // get channel name from discussion channel id
-    const { discussionChannelId, username } = args;
-
-    if (!discussionChannelId || !username) {
-      throw new Error(
-        "All arguments (discussionChannelId, username) are required"
-      );
-    }
-
-    const discussionChannelData = await DiscussionChannelModel.find({
-      where: { id: discussionChannelId },
-      selectionSet: `{ 
-        id
-        channelUniqueName
-      }`,
-    });
-
-    if (!discussionChannelData || !discussionChannelData[0]) {
-      throw new Error("No discussion channel found.");
-    }
-
-    const channelName = discussionChannelData[0]?.channelUniqueName;
-
-    if (!channelName) {
-      throw new Error("No channel found.");
-    }
-
-    const permissionResult = await hasChannelPermission({
-      permission: "canUpvoteDiscussion",
-      channelName,
-      context: ctx,
-    });
-
-    if (!permissionResult) {
-      throw new Error(ERROR_MESSAGES.channel.noChannelPermission);
-    }
-
-    if (permissionResult instanceof Error) {
-      return permissionResult;
-    }
-
-    return true;
-  }
-);
-
-const canGiveFeedback = rule({ cache: "contextual" })(
-  async (parent: unknown, args: unknown, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    const permissionResult = await hasServerModPermission(
-      "canGiveFeedback",
-      ctx
-    );
-
-    if (!permissionResult) {
-      return false;
-    }
-
-    if (permissionResult instanceof Error) {
-      return permissionResult;
-    }
-
-    return true;
-  }
-);
-
-const canReportContent = rule({ cache: "contextual" })(
-  async (parent: unknown, args: unknown, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    // Placeholder rule for now
-
-    return true;
-  }
-);
-
-const issueIsValid = rule({ cache: "contextual" })(
-  async (parent: unknown, args: unknown, ctx: GraphQLContext, info: GraphQLResolveInfo) => {
-    return true;
-  }
-);
+import {
+  canCreateChannel,
+  canCreateComment,
+  canCreateDiscussion,
+  canCreateEvent,
+} from "./definitions/contentCreationRules.js";
+import {
+  canEditWikiPages,
+  canDeleteWikiPages,
+  canEditWikiHomePage,
+} from "./definitions/wikiRules.js";
+import { canUpvoteComment, canUpvoteDiscussion } from "./definitions/votingRules.js";
+import {
+  isAdmin,
+  canUploadFile,
+  canGiveFeedback,
+  canReportContent,
+  issueIsValid,
+} from "./definitions/serverRules.js";
+
+// Re-exported for tests and the validation modules that import from "./rules".
+export { evaluateCanCreateChannelRule } from "./definitions/contentCreationRules.js";
+export type {
+  CreateDiscussionItem,
+  CanCreateDiscussionArgs,
+  CanUpdateDiscussionArgs,
+  SingleEventInput,
+  CanCreateEventArgs,
+} from "./definitions/contentCreationRules.js";
+export {
+  evaluateCanEditWikiPagesRule,
+  evaluateCanEditWikiHomePageRule,
+  evaluateCanDeleteWikiPagesRule,
+} from "./definitions/wikiRules.js";
 
 const ruleList = {
   isChannelOwner,
